move helper modules to marine_qc.helpers
