+++ src/marine_qc/__init__.py
@@ -2,7 +2,8 @@
 
 from __future__ import annotations
 
-from .auxiliary import (
+from .duplicate_checker.duplicates import flag_duplicates, remove_duplicates
+from .helpers.auxiliary import (
     PandasNAType,
     PandasNaTType,
     ScalarDatetimeType,
@@ -21,7 +22,14 @@ from .auxiliary import (
     ValueNumberType,
     ValueStrType,
 )
-from .duplicate_checker.duplicates import flag_duplicates, remove_duplicates
+from .helpers.external_clim import (
+    ClimArgType,
+    ClimFloatType,
+    ClimInputType,
+    ClimIntType,
+    ClimNumberType,
+    Climatology,
+)
 from .plotting.plot_qc_outcomes import (
     latitude_longitude_plot,
     latitude_variable_plot,
@@ -36,14 +44,6 @@ from .quality_control.buoy_tracking_qc import (
     do_sst_end_tail_check,
     do_sst_noisy_check,
     do_sst_start_tail_check,
-)
-from .quality_control.external_clim import (
-    ClimArgType,
-    ClimFloatType,
-    ClimInputType,
-    ClimIntType,
-    ClimNumberType,
-    Climatology,
 )
 from .quality_control.multiple_checks import do_multiple_grouped_check, do_multiple_individual_check, do_multiple_sequential_check
 from .quality_control.qc_grouped_reports import do_bayesian_buddy_check, do_mds_buddy_check
