+++ src/marine_qc/quality_control/buoy_tracking_qc.py
@@ -15,12 +15,12 @@ from datetime import datetime
 import numpy as np
 import pandas as pd
 
-from ..auxiliary import SequenceDatetimeType, SequenceNumberType, ensure_arrays, failed, inspect_arrays, isvalid, passed, untestable, untested
-from .astronomical_geometry import sunangle
+from ..helpers.astronomical_geometry import sunangle
+from ..helpers.auxiliary import SequenceDatetimeType, SequenceNumberType, ensure_arrays, failed, inspect_arrays, isvalid, passed, untestable, untested
+from ..helpers.spherical_geometry import sphere_distance
+from ..helpers.statistics import trim_mean, trim_std
+from ..helpers.time_control import convert_date_to_hours, day_in_year
 from .qc_sequential_reports import do_iquam_track_check
-from .spherical_geometry import sphere_distance
-from .statistics import trim_mean, trim_std
-from .time_control import convert_date_to_hours, day_in_year
 
 
 """
