+++ src/marine_qc/quality_control/multiple_checks.py
@@ -6,8 +6,8 @@ from typing import Any, Literal, cast
 
 import pandas as pd
 
-from ..auxiliary import failed, passed, untested
-from .external_clim import get_climatological_value  # noqa: F401
+from ..helpers.auxiliary import failed, passed, untested
+from ..helpers.external_clim import get_climatological_value  # noqa: F401
 from .qc_grouped_reports import (  # noqa: F401
     do_bayesian_buddy_check,
     do_mds_buddy_check,
