+++ src/marine_qc/quality_control/qc_grouped_reports.py
@@ -11,7 +11,7 @@ import math
 import numpy as np
 import pandas as pd
 
-from ..auxiliary import (
+from ..helpers.auxiliary import (
     SequenceDatetimeType,
     SequenceIntType,
     SequenceNumberType,
@@ -25,24 +25,24 @@ from ..auxiliary import (
     untestable,
     untested,
 )
-from .buoy_tracking_qc import is_monotonic
-from .external_clim import (
+from ..helpers.external_clim import (
     ClimArgType,
     Climatology,
     inspect_climatology,
 )
-from .location_control import (
+from ..helpers.location_control import (
     mds_lat_to_yindex,
     mds_lat_to_yindex_fast,
     mds_lon_to_xindex,
     mds_lon_to_xindex_fast,
 )
-from .statistics import p_gross
-from .time_control import (
+from ..helpers.statistics import p_gross
+from ..helpers.time_control import (
     convert_date,
     pentad_to_month_day,
     which_pentad,
 )
+from .buoy_tracking_qc import is_monotonic
 
 
 def get_threshold_multiplier(total_nobs: int, nob_limits: list[int], multiplier_values: list[float]) -> float:
