+++ src/marine_qc/quality_control/qc_individual_reports.py
@@ -9,7 +9,8 @@ from typing import Any, Literal
 
 import numpy as np
 
-from ..auxiliary import (
+from ..helpers.astronomical_geometry import sunangle
+from ..helpers.auxiliary import (
     ScalarNumberType,
     ValueDatetimeType,
     ValueFloatType,
@@ -25,9 +26,8 @@ from ..auxiliary import (
     post_format_return_type,
     untestable,
 )
-from .astronomical_geometry import sunangle
-from .external_clim import ClimArgType, inspect_climatology
-from .time_control import convert_date, day_in_year, get_month_lengths
+from ..helpers.external_clim import ClimArgType, inspect_climatology
+from ..helpers.time_control import convert_date, day_in_year, get_month_lengths
 
 
 vectorized_day_in_year = np.vectorize(day_in_year)
