+++ tests/test_calculate_humidity.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import numpy as np
 
-import marine_qc.quality_control.calculate_humidity as ch
+import marine_qc.helpers.calculate_humidity as ch
 
 
 def test_vap_from_example():
