+++ tests/test_individual_report_qc.py
@@ -24,7 +24,7 @@ from marine_qc import (
     do_valid_value_clim_check,
     do_wind_consistency_check,
 )
-from marine_qc.auxiliary import (
+from marine_qc.helpers.auxiliary import (
     convert_to,
     failed,
     passed,
