+++ tests/test_multiple_checks.py
@@ -10,7 +10,7 @@ from marine_qc import (
     do_multiple_individual_check,
     do_multiple_sequential_check,
 )
-from marine_qc.auxiliary import failed, passed, untested
+from marine_qc.helpers.auxiliary import failed, passed, untested
 from marine_qc.quality_control.multiple_checks import (
     _apply_qc_to_masked_rows,
     _do_multiple_check,
