+++ tests/test_track_check.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 import numpy as np
 import pytest
 
-import marine_qc.quality_control.spherical_geometry as sg
-from marine_qc.auxiliary import convert_to
+import marine_qc.helpers.spherical_geometry as sg
+from marine_qc.helpers.auxiliary import convert_to
 from marine_qc.quality_control.track_check_utils import (
     check_distance_from_estimate,
     direction_continuity,
