+++ tests/test_trackqc.py
@@ -8,7 +8,7 @@ import pandas as pd
 import pytest
 
 import marine_qc.quality_control.buoy_tracking_qc as tqc
-from marine_qc.auxiliary import failed, passed, untestable
+from marine_qc.helpers.auxiliary import failed, passed, untestable
 
 
 @pytest.mark.parametrize(
